DemandBusAdminWeb/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
Django==2.0.1
django-bootstrap-form==3.3
django-cors-headers==2.1.0
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
urllib3==1.22
